requirements.txt
# Movie Insights - Dependencies

# Scene Detection
scenedetect[opencv]>=0.6.2

# Video Processing
opencv-python>=4.8.0

# Excel Export
openpyxl>=3.1.2
Pillow>=10.0.0

# PowerPoint Export
python-pptx>=0.6.21

# Web UI
streamlit>=1.28.0

# CLI
click>=8.1.0
